extract CCIP status component
